Refactors color module structure for improved organization
Renames the color-related file to a more general "index" to simplify imports and improve module organization. Updates import paths accordingly to use the new structure, reducing redundancy and enhancing maintainability.
