+++ src/lint/colors/index.ts
